ZDBSP: hack together some basic functionality
- This is very much not up to the standards for a public interface that I'd be willing to export, but it's at least usable.
- Move off of cxx - which has proven too inflexible for these purposes - to a combination of bindgen and cbindgen.
- The Rust API will now be complemented by a C API rather than a C++ API.

+++ zdbsp/src/blockmapbuilder.cpp
@@ -72,7 +72,7 @@ int BoxOnSide (int bx1, int by1, int bx2, int by2,
 }
 #endif
 
-WORD *FBlockmapBuilder::GetBlockmap (int &size)
+WORD *FBlockmapBuilder::GetBlockmap (size_t &size)
 {
 #ifdef BLOCK_TEST
 	FILE *f = fopen ("blockmap.lmp", "rb");
